+++ packages/next-common/utils/consts/settings/litmus.js
@@ -1,4 +1,5 @@
 import Logo from "../../../assets/header-logos/litmus.svg";
+import DarkModeLogo from "../../../assets/header-logos/litmus-dark.svg";
 import Avatar from "../../../assets/icons/chain/litmus.png";
 
 import Chains from "../chains";
@@ -24,9 +25,8 @@ const litmus = {
   snsCoverCid: "bafybeiexfrkdte6eruqghlc66xpnfoyadkgq5we3ql5elqjwgdzbxmez6q",
   endpoints: DEFAULT_LITMUS_NODES,
   loginButtonPrimary: true,
-  headerBackgroundColor: "#000000",
   headerLogo: Logo,
-  darkHeaderLogo: Logo,
+  darkHeaderLogo: DarkModeLogo,
   avatar: Avatar,
   group: MenuGroups.KusamaAndParachains,
   hasStatescan: true,